+++ src/rename-file/rename-file.js
@@ -1,10 +1,10 @@
 const { sep } = require("path")
 
-const trim = require("../string__trim/trim")
-const split = require("../string__split/split")
-const push = require("../array__push/push")
-const join = require("../array__join/join")
-const dropLast = require("../array__drop-last/drop-last")
+const trim = require("../trim/trim")
+const split = require("../split/split")
+const push = require("../push/push")
+const join = require("../join/join")
+const dropLast = require("../drop-last/drop-last")
 
 const removeTrailingSlash = source =>
   source[source.length - 1] === sep ? source.slice(0, -1) : source